requirements.txt
Pillow==8.4.0
PyAutoGUI==0.9.53
